refactor(repositories): fix repositories path
